-- Fixed the updateGameStatus() "tallying" section to look for a streak of >= mMovesNeededToWin *2 It was previously considering a streak of equal size a winner and this was causing less moves to win
-- Added the the checkVerical iteration sequence

All seems to be in order, I have tested the ability to win on a row and column in all game modes and it seems to be responding as expected.

Note there are still some instances where we see an erroneous tie but I think it will self correct once we add in the check diagonal.